requirements.txt
flask==3.0.0
flask-cors==4.0.0
SpeechRecognition==3.10.0
pyttsx3==2.90
pywhatkit==5.4
wikipedia==1.4.0
pyjokes==0.6.0
requests==2.31.0
python-dotenv==1.0.0
PyAudio==0.2.14
